Programa produtor-consumidor funcionando

+++ programaca-concorrente-2024-01/lab6/atividade2/src/include/ehprimo.c
@@ -6,7 +6,7 @@ int ehPrimo(long long int n) {
     if (n==2) return 1;
     if (n%2==0) return 0;
     
-    for (i=3; i<sqrt(n)+1; i+=2)
+    for (i=3; i<sqrt((double) n)+1; i+=2)
         if(n%i==0) return 0;
 
     return 1;
